✨ Wire up a thin Continuity protocol to the stack
Currently, the protocol does nothing, but some ideas are logged to the
README that sits alongside it.

The gist:

The Continuity protocol layer provides sequencing to messages, allowing
messages received out of sequence to be reconstructed or discarded, or
duplicate messages to be discarded, or for messages which fail to include
the hash of the previous message to be discarded or flagged. This provides
some degree of session integrity even if a user's private key is
compromised.